완료한 것  - 평점 주기 기능추가. 해야할 것  - 댓글좋아요, 댓글 싫어요  - 푸쉬

+++ MyNaverWebtoon3/DataClass/ComicContentsDTO.swift
@@ -21,7 +21,7 @@ class ComicContentsDTO: BaseDTO {
         var contentDate:String?
         var contentHeart:Int?
         var contentType:Int?
-        var contentRating:Int?
+        var contentRating:Float?
         required init?(map: Map){
             
         }
